Moved ImageChoiceField to admin.py
